test(switchyard): move contract tests to integration suite
